fix: ensured to populate correct js file
