+++ src/Our.Umbraco.ValidationAttributes.StaticAssets/StaticAssetsBoot.cs
@@ -1,0 +1,12 @@
+using Umbraco.Cms.Core.Composing;
+using Umbraco.Cms.Core.DependencyInjection;
+
+namespace Our.Umbraco.ValidationAttributes.UI;
+
+public class StaticAssetsBoot : IComposer
+{
+    public void Compose(IUmbracoBuilder builder)
+    {
+        builder.AddValidationAttributesStaticAssets();
+    }
+}
